Revert "Fixed Minor Code Errors"
This reverts commit 788c42e2fa43d2317ac87168ae701a2bdc29116d.

+++ Assets/scripts/BaseStat.cs
@@ -5,9 +5,9 @@ public class BaseStat {
 
 	private string name; //name of stat
 	private string description; //short description
-	private float baseValue; //base stat value
-	private float modifiedValue; //modifier ex. certain weps or pots
-	private float maxValue; //max value
+	private double baseValue; //base stat value
+	private double modifiedValue; //modifier ex. certain weps or pots
+	private double maxValue; //max value
 	private StatTypes type; //stat type
 
 	public enum StatTypes{
@@ -33,21 +33,21 @@ public class BaseStat {
 		
 	}
 
-	public float StatBaseValue {
+	public double StatBaseValue {
 		
 		get{ return baseValue; }
 		set{ baseValue = value; }
 		
 	}
 
-	public float StatModifiedValue {
+	public double StatModifiedValue {
 		
 		get{ return modifiedValue; }
 		set{ modifiedValue = value; }
 		
 	}
 
-	public float StatMaxValue {
+	public double StatMaxValue {
 		
 		get{ return maxValue; }
 		set{ maxValue = value; }
